update/ adjust SignUpForm Component
